+++ snapatac2-python/python/snapatac2/preprocessing/_basic.py
@@ -362,7 +362,7 @@ def add_tile_matrix(
     exclude_chroms: list[str] | str | None = ["chrM", "chrY", "M", "Y"],
     min_frag_size: int | None = None,
     max_frag_size: int | None = None,
-    count_frag_as_reads: bool = True,
+    counting_strategy: Literal['fragment', 'insertion', 'paired-insertion'] = 'insertion',
     file: Path | None = None,
     backend: Literal['hdf5'] = 'hdf5',
     n_jobs: int = 8,
@@ -393,9 +393,16 @@ def add_tile_matrix(
         Minimum fragment size to include.
     max_frag_size
         Maximum fragment size to include.
-    count_frag_as_reads
-        Whether to count fragments as reads. If `True`, each fragment is converted
-        to two points representing both ends of the fragment.
+    counting_strategy
+        The strategy to compute feature counts. It must be one of the following:
+        "fragment", "insertion", or "paired-insertion". "fragment" means the
+        feature counts are assigned based on the number of fragments that overlap
+        with a region of interest. "insertion" means the feature counts are assigned
+        based on the number of insertions that overlap with a region of interest.
+        "paired-insertion" is similar to "insertion", but it only counts the insertions
+        once if the pair of insertions of a fragment are both within the same region
+        of interest [Miao24]_.
+        Note that this parameter has no effect if input are single-end reads.
     file
         File name of the output file used to store the result. If provided, result will
         be saved to a backed AnnData, otherwise an in-memory AnnData is used.
@@ -436,11 +443,11 @@ def add_tile_matrix(
         if isinstance(adata, list):
             snapatac2._utils.anndata_par(
                 adata,
-                lambda x: internal.mk_tile_matrix(x, bin_size, chunk_size, count_frag_as_reads, exclude_chroms, None),
+                lambda x: internal.mk_tile_matrix(x, bin_size, chunk_size, counting_strategy, exclude_chroms, None),
                 n_jobs=n_jobs,
             )
         else:
-            internal.mk_tile_matrix(adata, bin_size, chunk_size, count_frag_as_reads, exclude_chroms, min_frag_size, max_frag_size, None)
+            internal.mk_tile_matrix(adata, bin_size, chunk_size, counting_strategy, exclude_chroms, min_frag_size, max_frag_size, None)
     else:
         if file is None:
             if adata.isbacked:
@@ -449,7 +456,7 @@ def add_tile_matrix(
                 out = AnnData(obs=adata.obs[:])
         else:
             out = internal.AnnData(filename=file, backend=backend, obs=adata.obs[:])
-        internal.mk_tile_matrix(adata, bin_size, chunk_size, count_frag_as_reads, exclude_chroms, min_frag_size, max_frag_size, out)
+        internal.mk_tile_matrix(adata, bin_size, chunk_size, counting_strategy, exclude_chroms, min_frag_size, max_frag_size, out)
         return out
 
 def make_peak_matrix(
@@ -464,7 +471,7 @@ def make_peak_matrix(
     use_x: bool = False,
     min_frag_size: int | None = None,
     max_frag_size: int | None = None,
-    count_frag_as_reads: bool = True,
+    counting_strategy: Literal['fragment', 'insertion', 'paired-insertion'] = 'insertion',
 ) -> internal.AnnData:
     """Generate cell by peak count matrix.
 
@@ -502,9 +509,16 @@ def make_peak_matrix(
         Minimum fragment size to include.
     max_frag_size
         Maximum fragment size to include.
-    count_frag_as_reads
-        Whether to count fragments as reads. If `True`, each fragment is converted
-        to two points representing both ends of the fragment.
+    counting_strategy
+        The strategy to compute feature counts. It must be one of the following:
+        "fragment", "insertion", or "paired-insertion". "fragment" means the
+        feature counts are assigned based on the number of fragments that overlap
+        with a region of interest. "insertion" means the feature counts are assigned
+        based on the number of insertions that overlap with a region of interest.
+        "paired-insertion" is similar to "insertion", but it only counts the insertions
+        once if the pair of insertions of a fragment are both within the same region
+        of interest [Miao24]_.
+        Note that this parameter has no effect if input are single-end reads.
 
     Returns
     -------
@@ -558,7 +572,7 @@ def make_peak_matrix(
             out = AnnData(obs=adata.obs[:])
     else:
         out = internal.AnnData(filename=file, backend=backend, obs=adata.obs[:])
-    internal.mk_peak_matrix(adata, peaks, chunk_size, use_x, count_frag_as_reads, min_frag_size, max_frag_size, out)
+    internal.mk_peak_matrix(adata, peaks, chunk_size, use_x, counting_strategy, min_frag_size, max_frag_size, out)
     return out
 
 def make_gene_matrix(
@@ -577,7 +591,7 @@ def make_gene_matrix(
     gene_id_key: str = "gene_id",
     min_frag_size: int | None = None,
     max_frag_size: int | None = None,
-    count_frag_as_reads: bool = True,
+    counting_strategy: Literal['fragment', 'insertion', 'paired-insertion'] = 'insertion',
 ) -> internal.AnnData:
     """Generate cell by gene activity matrix.
 
@@ -619,9 +633,16 @@ def make_gene_matrix(
         Minimum fragment size to include.
     max_frag_size
         Maximum fragment size to include.
-    count_frag_as_reads
-        Whether to count fragments as reads. If `True`, each fragment is converted
-        to two points representing both ends of the fragment.
+    counting_strategy
+        The strategy to compute feature counts. It must be one of the following:
+        "fragment", "insertion", or "paired-insertion". "fragment" means the
+        feature counts are assigned based on the number of fragments that overlap
+        with a region of interest. "insertion" means the feature counts are assigned
+        based on the number of insertions that overlap with a region of interest.
+        "paired-insertion" is similar to "insertion", but it only counts the insertions
+        once if the pair of insertions of a fragment are both within the same region
+        of interest [Miao24]_.
+        Note that this parameter has no effect if input are single-end reads.
 
     Returns
     -------
@@ -658,7 +679,7 @@ def make_gene_matrix(
         out = internal.AnnData(filename=file, backend=backend, obs=adata.obs[:])
     internal.mk_gene_matrix(adata, gene_anno, chunk_size, use_x, id_type,
         transcript_name_key, transcript_id_key, gene_name_key, gene_id_key,
-        count_frag_as_reads, min_frag_size, max_frag_size, out)
+        counting_strategy, min_frag_size, max_frag_size, out)
     return out
 
 def filter_cells(